fix lint feedback

+++ backup-docker.sh
@@ -31,8 +31,8 @@ fi
 HOSTNAME="$(hostname)"
 TIMESTAMP=$(date +"%Y%m%d_%H%M")
 BACKUPDIR="${2:-"/mnt/backup"}/${HOSTNAME}"
-if [ ! -d $BACKUPDIR ]; then
-	mkdir -p $BACKUPDIR
+if [ ! -d "$BACKUPDIR" ]; then
+	mkdir -p "$BACKUPDIR"
 fi
 TEMPDIR="/var/tmp"
 BACKUPDURATIONDAYS=${3:-2}
@@ -40,8 +40,8 @@ DOCKERROOTDIR=$(docker info --format '{{ .DockerRootDir }}')
 # get docker-compose project name from variable or from current directory (lower case)
 PROJECTNAME=$(echo "${1:-$(basename "$PWD")}" | tr '[:upper:]' '[:lower:]')
 PROJECTBACKUPDIR="${BACKUPDIR}/${PROJECTNAME}"
-if [ ! -d $PROJECTBACKUPDIR ]; then
-	mkdir -p $PROJECTBACKUPDIR
+if [ ! -d "$PROJECTBACKUPDIR" ]; then
+	mkdir -p "$PROJECTBACKUPDIR"
 fi
 ALLCONTAINER=$(docker ps -q --filter "label=com.docker.compose.project=$PROJECTNAME")
 if [ -z "$ALLCONTAINER" ]; then
@@ -51,7 +51,7 @@ fi
 WORKINGDIR=$(for i in $ALLCONTAINER; do
   docker inspect --format '{{ index .Config.Labels "com.docker.compose.project.working_dir"}}' "$i"
 done | sort -u | head -n 1)
-cd $WORKINGDIR
+cd "$WORKINGDIR"
 # =======================================================================
 # PUT FUNCTIONS HERE
 # Print Status
@@ -75,7 +75,7 @@ START=$(date +%s)
 echo "Backup Docker Project: ${PROJECTNAME}"
 print_status "  📦 DockerCompose: $WORKINGDIR... "
 OUTPUT=${TIMESTAMP}_${PROJECTNAME}.compose.tar.gz
-tar -czf ${TEMPDIR}/${OUTPUT} -C $WORKINGDIR .
+tar -czf ${TEMPDIR}/${OUTPUT} -C "$WORKINGDIR" .
 cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 END=$(date +%s); DURATION=$((END - START))
 print_duration $DURATION
@@ -95,24 +95,24 @@ for cont in $CONTAINERS; do
 			# BACKUP MariaDB
 			print_status "  🐬 MariaDB: ${PROJECTNAME}.${CONTAINERNAME}.mariadbdump.sql.gz... "
 			OUTPUT=${TIMESTAMP}_${PROJECTNAME}.${CONTAINERNAME}.mariadbdump.sql.gz
-			CONTAINERENV_DBPW=$(docker exec ${cont} sh -c 'echo "$MYSQL_ROOT_PASSWORD $DB_ROOT_PASSWORD"')
-			docker exec ${cont} sh -c 'exec mariadb-dump -u root -p"$0" --all-databases 1>/dev/null' ${CONTAINERENV_DBPW} | gzip -9 > ${TEMPDIR}/${OUTPUT}
+			CONTAINERENV_DBPW=$(docker exec "${cont}" sh -c 'echo "$MYSQL_ROOT_PASSWORD $DB_ROOT_PASSWORD"')
+			docker exec "${cont}" sh -c 'exec mariadb-dump -u root -p"$0" --all-databases 1>/dev/null' "${CONTAINERENV_DBPW}" | gzip -9 > "${TEMPDIR}/${OUTPUT}"
 			cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 		elif echo "$IMAGE" | grep -qi "mysql"; then
 			# =======================================================================
 			# BACKUP MySQL
 			print_status "  🐬 MySQL: ${PROJECTNAME}.${CONTAINERNAME}.mysqldump.sql.gz... "
 			OUTPUT=${TIMESTAMP}_${PROJECTNAME}.${CONTAINERNAME}.mysqldump.sql.gz
-			CONTAINERENV_DBPW=$(docker exec ${cont} sh -c 'echo "$MYSQL_ROOT_PASSWORD $DB_ROOT_PASSWORD"')
-			docker exec ${cont} sh -c 'exec mysqldump -u root -p"$0" --all-databases 1>/dev/null' ${CONTAINERENV_DBPW} | gzip -9 > ${TEMPDIR}/${OUTPUT}
+			CONTAINERENV_DBPW=$(docker exec "${cont}" sh -c 'echo "$MYSQL_ROOT_PASSWORD $DB_ROOT_PASSWORD"')
+			docker exec "${cont}" sh -c 'exec mysqldump -u root -p"$0" --all-databases 1>/dev/null' "${CONTAINERENV_DBPW}" | gzip -9 > "${TEMPDIR}/${OUTPUT}"
 			cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 		elif echo "$IMAGE" | grep -qi "postgres"; then
 			# =======================================================================
 			# BACKUP PostgreSQL
 			print_status "  🐘️ PostgreSQL: ${PROJECTNAME}.${CONTAINERNAME}.postgredump.sql.gz... "
 			OUTPUT=${TIMESTAMP}_${PROJECTNAME}.${CONTAINERNAME}.postgredump.sql.gz
-			CONTAINERENV_DBUSER=$(docker exec ${cont} sh -c 'echo "$POSTGRES_USER"')
-			docker exec ${cont} sh -c 'exec pg_dumpall -U "$0"' ${CONTAINERENV_DBUSER} | gzip -9 > ${TEMPDIR}/${OUTPUT}
+			CONTAINERENV_DBUSER=$(docker exec "${cont}" sh -c 'echo "$POSTGRES_USER"')
+			docker exec "${cont}" sh -c 'exec pg_dumpall -U "$0"' "${CONTAINERENV_DBUSER}" | gzip -9 > "${TEMPDIR}/${OUTPUT}"
 			cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 		elif echo "$IMAGE" | grep -qi "mongo"; then
 			# =======================================================================
@@ -120,14 +120,14 @@ for cont in $CONTAINERS; do
 			if echo "$vol" | grep -qi "config"; then
 				VOLUMEDIR=${DOCKERROOTDIR}/volumes/$vol
 				print_status "  💾 LocalStorage: ${PROJECTNAME}.${VOLUMENAME}.volume.tar.gz... "
-				cd $VOLUMEDIR
+				cd "$VOLUMEDIR"
 				OUTPUT=${TIMESTAMP}_${PROJECTNAME}.${VOLUMENAME}.volume.tar.gz
-				tar -czf ${TEMPDIR}/${OUTPUT} .
+				tar -czf "${TEMPDIR}/${OUTPUT}" .
 				cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 			else
 				print_status "  🍃 MongoDB: ${PROJECTNAME}.${CONTAINERNAME}.mongodump.sql.gz... "
 				OUTPUT=${TIMESTAMP}_${PROJECTNAME}.${CONTAINERNAME}.mongodump.sql.gz
-				docker exec ${cont} sh -c 'mongodump --archive --gzip --quiet' > ${TEMPDIR}/${OUTPUT}
+				docker exec ${cont} sh -c 'mongodump --archive --gzip --quiet' > "${TEMPDIR}/${OUTPUT}"
 				cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 			fi
 		elif echo "$IMAGE" | grep -qiE "^gitlab/gitlab"; then
@@ -137,7 +137,7 @@ for cont in $CONTAINERS; do
 			OUTPUT=${TIMESTAMP}_${PROJECTNAME}.${CONTAINERNAME}.gitlabbackup.tar.gz
 			docker exec -u root ${cont} bash -c "mkdir -p /mnt/backup-cache/code && chown git /mnt/backup-cache/code && chmod 700 /mnt/backup-cache/code"
 			docker exec -u root ${cont} bash -c "export COMPRESS_CMD=gzip SKIP=artifacts,registry && gitlab-backup create --quiet"			
-			tar -czf ${TEMPDIR}/${OUTPUT} -C /mnt/backup-cache/gitlab-backup .
+			tar -czf "${TEMPDIR}/${OUTPUT}" -C /mnt/backup-cache/gitlab-backup .
 			# rm -rf /mnt/backup-cache/gitlab-backup/code
 			cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 			# Cancel other volume backups for this container
@@ -151,8 +151,8 @@ for cont in $CONTAINERS; do
 			printf "%-100s" "  💾 LocalStorage: ${PROJECTNAME}.${VOLUMENAME}.volume.tar.gz... "
 			OUTPUT=${TIMESTAMP}_${PROJECTNAME}.${VOLUMENAME}.volume.tar.gz
 			cp -r "${VOLUMEDIR}" "${TEMPDIR}"
-			tar -czf ${TEMPDIR}/${OUTPUT} -C "${TEMPDIR}/${vol}" .
-			rm -r "${TEMPDIR}/${vol}"
+			tar -czf "${TEMPDIR}/${OUTPUT}" -C "${TEMPDIR}/${vol}" .
+			rm -r "${TEMPDIR:?}/${vol:?}"
 			cp "${TEMPDIR}/${OUTPUT}" "${PROJECTBACKUPDIR}" && rm "${TEMPDIR}/${OUTPUT}"
 		fi
 		END=$(date +%s); DURATION=$((END - START))
@@ -161,7 +161,7 @@ for cont in $CONTAINERS; do
 done
 # =======================================================================
 echo "  🗑️$ Cleanup old backups (older than ${BACKUPDURATIONDAYS} days)..."
-find "$PROJECTBACKUPDIR" -name "*_$PROJECTNAME*.gz" -daystart -mtime +$BACKUPDURATIONDAYS | while read -r file; do
+find "$PROJECTBACKUPDIR" -name "*_$PROJECTNAME*.gz" -daystart -mtime +"$BACKUPDURATIONDAYS" | while read -r file; do
   echo "    - Delete: $file"
   rm "$file"
 done
